Revert "Feature/rsi service (#20)"
This reverts commit 7b2114940461867ce2ef0ff54fdbe5ab0cafbcf3.
