--- api/rsi.proto
@@ -1,20 +1,0 @@
-syntax = "proto3";
-import "base.proto";
-
-package rsi;
-
-option go_package = "internal/proto/monitor";
-
-service Rsi {
-  rpc GetRsi (PriceData) returns (RsiData);
-}
-
-message PriceData {
-  repeated string pd = 1;
-  base.RequestInfo debug = 2;
-}
-
-message RsiData {
-  string rsival = 1;
-  base.RequestInfo debug = 2;
-}
