+++ rsi_service/Cargo.toml
@@ -9,12 +9,11 @@ edition = "2021"
 tonic = "0.7"
 prost = "0.10"
 prost-types = "0.10"
-rust_decimal_macros = "1.25"
-tokio-test = "0.4"
-rust_decimal = { version = "1.25", features = ["maths"] }
-sqlx = { version = "0.6", features = ["mysql", "runtime-tokio-native-tls"] }
-tokio = { version = "1.20", features = ["macros", "rt-multi-thread"] }
-uuid = { version = "1.1", features = ["v4"] }
+tokio = { version = "1.0", features = ["macros", "rt-multi-thread"] }
+
+[[bin]] # Bin to run the HelloWorld gRPC server
+name = "rsi-grpc-server"
+path = "src/server.rs"
 
 [build-dependencies]
 tonic-build = "0.7"
