+++ rsi_service/build.rs
@@ -1,8 +1,7 @@
 fn main() -> Result<(), Box<dyn std::error::Error>> {
    tonic_build::configure()
-        .build_client(false)
-        .out_dir("./src/protos/")
-        .compile(&["../api/monitor.proto", "../api/rsi.proto"],
+        .out_dir("src/protos/")
+        .compile(&["../api/monitor.proto"],
                  &["../api/"])?;
    Ok(())
 }
