--- rsi_service/src/protos.rs
@@ -1,3 +1,0 @@
-pub mod monitor;
-pub mod rsi;
-pub mod base;
